requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
aiofiles==0.7.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
cloudflare==2.8.15
paramiko==2.8.1
PyYAML==6.0
python-dotenv==0.19.0
